requirements.txt
yfinance>=0.2.36
openai>=1.12.0
pyyaml>=6.0
pandas>=2.0.0
numpy>=1.24.0
requests>=2.31.0
beautifulsoup4>=4.12.0
selenium>=4.15.0
sqlalchemy>=2.0.0
alembic>=1.13.0
feedparser>=6.0.0
sec-edgar-downloader>=5.0.0
python-dotenv>=1.0.0
chromadb>=0.4.0
tiktoken>=0.5.0
schedule>=1.2.0
aiohttp>=3.9.0
trafilatura>=1.6.0
